specify that whitespace should be stripped before checking the header
Now that we have inline headers and footers, it's possible that an
application that inserts newlines in our messages will put one right in
the middle of the footer. (The same might've been a problem even before
inline headers, e.g. for very small column widths.) If we require that
words in the header and footer are separated by exactly one space, we're
going to break on this case.

The simplest fix seems to be stripping whitespace first, and then
matching against an expression that has no spaces. I don't think there's
any rush to change our implementation, since this is unlikely to come up
in practice, but it would be nice to spec the right thing.